requirements.txt
Flask
flask-cors
requests
google-genai
ddgs
PyMuPDF
python-dotenv
gunicorn
fastapi